Extract Tape to lib module
