+++ settings.gradle.kts
@@ -27,6 +27,7 @@ subproject(name = "app", path = file("subprojects/app"))
 subproject(name = "feature-repos", path = file("subprojects/feature/repos"))
 subproject(name = "feature-details", path = file("subprojects/feature/details"))
 subproject(name = "retrofit-cache", path = file("subprojects/libs/retrofit-cache"))
+subproject(name = "test-tape", path = file("subprojects/libs/test-tape"))
 
 /**
  * Helper function that uses name as artifact name and accepts the relative path to point out the sources location.
