+++ subprojects/feature/repos/src/test/kotlin/com/olderwold/jlabs/github/feature/repos/data/GithubApiTest.kt
@@ -1,5 +1,6 @@
 package com.olderwold.jlabs.github.feature.repos.data
 
+import com.olderwold.jlabs.github.tape.tape
 import kotlinx.coroutines.runBlocking
 import okreplay.OkReplay
 import org.amshove.kluent.shouldNotBeEmpty
