+++ subprojects/libs/test-tape/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("build.logic.jvm")
+    id("build.logic.kotlin")
+    id("build.logic.kotlin.checks")
+}
+
+dependencies {
+    implementation(testLibs.junit4)
+    implementation(testLibs.bundles.network)
+}
